sort scores
