+++ src/apps.rs
@@ -18,6 +18,12 @@ pub struct App {
     pub scores: Option<Vec<ScoreInfo>>,
 }
 
+impl Gt for App {
+    fn gt(&self, other: &Self) -> bool {
+        ascii_gt(&self.name, &other.name)
+    }
+}
+
 impl App {
     pub fn try_load_stats(&mut self) {
         if self.stats.is_some() {
@@ -77,31 +83,4 @@ pub fn read_apps() -> Vec<App> {
     }
     bubble_sort(&mut apps);
     apps
-}
-
-/// Good old bubble sort. Slower but much smaller than the built-in sort function.
-fn bubble_sort(apps: &mut [App]) {
-    let len = apps.len();
-    let mut sorted = false;
-    while !sorted {
-        sorted = true;
-        for i in 0..len - 1 {
-            if ascii_gt(&apps[i].name, &apps[i + 1].name) {
-                apps.swap(i, i + 1);
-                sorted = false;
-            }
-        }
-    }
-}
-
-/// Case-insensitive comparison of two ASCII strings.
-fn ascii_gt(s1: &str, s2: &str) -> bool {
-    for (c1, c2) in s1.as_bytes().iter().zip(s2.as_bytes()) {
-        let c1 = c1.to_ascii_lowercase();
-        let c2 = c2.to_ascii_lowercase();
-        if c1 != c2 {
-            return c1 > c2;
-        }
-    }
-    s1.len() > s2.len()
 }
